flytter logikk til tjeneste som skal gjenbrukes
